# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WhatsApp/NFS-e API",
        "description": "API de integração entre agendamentos e providers WhatsApp/NFS-e.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor atual"
        }
    ],
    "paths": {
...
        "/api/v1/whatsapp/senders": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Lista senders (número por estabelecimento)",
                "operationId": "f90abb9a2e57d5a0367af345340ddfce",
                "parameters": [
                    {
                        "name": "X-Client-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Nonce",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Signature",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de senders"
                    },
                    "401": {
                        "description": "Assinatura inválida"
                    }
                },
                "security": [
                    {
                        "hmac": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Cria/atualiza o sender de um estabelecimento",
                "operationId": "b308ad9db700689974743d5c9b43bdf6",
                "parameters": [
                    {
                        "name": "X-Client-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Timestamp",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Nonce",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Signature",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "establishment_external_id",
                                    "phone_number_id",
                                    "access_token"
                                ],
                                "properties": {
                                    "establishment_external_id": {
                                        "type": "string",
                                        "example": "est-001"
                                    },
                                    "phone_number_id": {
                                        "type": "string",
                                        "example": "1314669508391450"
                                    },
                                    "waba_id": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "display_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "access_token": {
                                        "description": "Token do estabelecimento (armazenado criptografado)",
                                        "type": "string"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "default": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sender salvo"
                    },
                    "401": {
                        "description": "Assinatura inválida"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "hmac": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "hmac": {
                "type": "apiKey",
                "name": "X-Signature",
                "in": "header"
            }
        }
    },
...
}